Get rid of a compiler warning
Commit e7d0e97bdbdc996564f06b382af3d5a5164a8fb3 ('hostapd: Add vendor
specific VHT extension for the 2.4 GHz band') resulted in a compiler
warning regarding comparison between signed and unsigned integers at
least for 32-bit builds.

+++ src/ap/ieee802_11_vht.c
@@ -134,7 +134,7 @@ u16 copy_sta_vendor_vht(struct hostapd_data *hapd, struct sta_info *sta,
 		goto no_capab;
 	vht_capab_len = vht_capab[1];
 	if (vht_capab_len < sizeof(struct ieee80211_vht_capabilities) ||
-	    vht_capab_len > ie + len - vht_capab - 2)
+	    (int) vht_capab_len > ie + len - vht_capab - 2)
 		goto no_capab;
 	vht_capab += 2;
